move websocket reducers to websocket dir
